Removed custom hook and used useContext
