+++ src/components/Modal/Modal.scss
@@ -28,4 +28,15 @@
       color: #ff004c;
     }
   }
+
+  &__button {
+    padding: 10px;
+    border: none;
+    border-radius: 10px;
+    background-color: #818384;
+    color: #fff;
+    font-weight: bold;
+    letter-spacing: 1px;
+    cursor: pointer;
+  }
 }
